+++ christmas_app/accounts.py
@@ -1,1 +1,11 @@
 """ Used to pre-defined user accounts """
+from models import User
+
+def create_accounts():
+    create_dad()
+
+def create_dad():
+    try:
+        pass
+    except:
+        pass