+++ changelog.md
@@ -17,3 +17,5 @@ Config the System version in **init**.py
 - Initial Development#2.2: comprehensive application draft implemented on December 4, 2022 -> **0.22**
 
 - Initial Development#3: offline cdn implemented with SweetAlert2 tested on December 7, 2022 -> **0.221**
+
+- Initial Development#4: login page and pre-account draft implemented on December 9, 2022 -> **0.23**